impl: MoveActor main collision sweep/response loop (Ghidra 0x103b9750)
Replaces the placeholder stub with a full Unreal-style collision implementation:

- MultiLineCheck sweep with correct TraceFlags (bCollideActors/World/Pawns)
- MoveActorFirstBlocking to find the first truly blocking hit
- Inlined FUN_103b7390 (touch collector): gathers non-blocking actors
  with Hit.Time < blocking Hit.Time into a local TouchList[256]
- Partial-delta trimming with 2-unit wall-backoff margin
  (dist = (Size+2)*HitTime; adj = Dir*(dist-2); HitTime = (dist-2)/Size)
- IsMovingBrush check (vtable+0xC8): routes movers through CheckEncroachment
- Hash Remove/Add around location update
- Old-rotation save for attached-actor delta-rotation detection
- UpdateRelativeRotation fallback for attached actors (TODO: full matrix loop)
- vtable+0x140 physics-sync call (gates proprietary Karma updates, omitted)
- vtable+0xcc bump notifications on both parties when blocking actor found
- vtable+0x70 / vtable+0xc4 touch-notify loop for non-blocking sweep contacts
- EndTouch loop for Touching[] actors no longer overlapping
- vtable+0x10c post-move callback
- FMemMark pop + UpdateRenderData at exit

Permanent divergences documented:
- Attached actor loop uses UpdateRelativeRotation instead of full matrix product
- AR6ColBox base step-up (AR6ColBox::GetMaxStepUp) omitted (R6-specific)
- Karma sync (FUN_104c3660 / KU2METransform / FUN_104aa490) omitted (proprietary)

+++ src/Engine/Src/UnLevel.cpp
@@ -1512,17 +1512,17 @@ if ( (ActorFlags & 0x800) && Hash )
 Hash->RemoveActor(Actor);
 
 // Save old rotation for delta-rotation computation for attached actors
-UINT OldPitch = *(UINT*)((BYTE*)Actor + 0x240);
-UINT OldYaw   = *(UINT*)((BYTE*)Actor + 0x244);
-UINT OldRoll  = *(UINT*)((BYTE*)Actor + 0x248);
+DWORD OldPitch = *(DWORD*)((BYTE*)Actor + 0x240);
+DWORD OldYaw   = *(DWORD*)((BYTE*)Actor + 0x244);
+DWORD OldRoll  = *(DWORD*)((BYTE*)Actor + 0x248);
 
 // Apply movement and new rotation
 *(FLOAT*)((BYTE*)Actor + 0x234) += AdjustedDelta.X;
 *(FLOAT*)((BYTE*)Actor + 0x238) += AdjustedDelta.Y;
 *(FLOAT*)((BYTE*)Actor + 0x23C) += AdjustedDelta.Z;
-*(UINT*)((BYTE*)Actor  + 0x240) = (UINT)NewRotation.Pitch;
-*(UINT*)((BYTE*)Actor  + 0x244) = (UINT)NewRotation.Yaw;
-*(UINT*)((BYTE*)Actor  + 0x248) = (UINT)NewRotation.Roll;
+*(INT*)((BYTE*)Actor  + 0x240) = NewRotation.Pitch;
+*(INT*)((BYTE*)Actor  + 0x244) = NewRotation.Yaw;
+*(INT*)((BYTE*)Actor  + 0x248) = NewRotation.Roll;
 
 // Propagate movement/rotation to attached actors.
 // Ghidra builds a delta-rotation matrix (GetLocalCoords vtable+0xa8,
@@ -1532,9 +1532,9 @@ UINT OldRoll  = *(UINT*)((BYTE*)Actor + 0x248);
 INT numAttached = *(INT*)((BYTE*)Actor + 0x1D8);  // Attached.Num()
 if ( numAttached > 0 && bTest == 0 )
 {
-UBOOL bRotChanged = ( OldPitch != (UINT)NewRotation.Pitch
-                   || OldYaw   != (UINT)NewRotation.Yaw
-                   || OldRoll  != (UINT)NewRotation.Roll );
+UBOOL bRotChanged = ( OldPitch != (DWORD)NewRotation.Pitch
+                   || OldYaw   != (DWORD)NewRotation.Yaw
+                   || OldRoll  != (DWORD)NewRotation.Roll );
 if ( bRotChanged )
 Actor->UpdateRelativeRotation();
 }
